requirements.txt
alembic==1.13.2
audioread==3.0.1
av==12.3.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
fast-bss-eval==0.1.4
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
future==1.0.0
greenlet==3.0.3
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mir-eval==0.7
moviepy==1.0.3
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
optuna==3.6.1
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.0
pooch==1.8.2
proglog==0.1.10
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sns==0.1
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.31
statsmodels==0.14.2
sympy==1.12.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.7.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
